+++ CHANGELOG.md
@@ -5,6 +5,10 @@ FEATURES
 * Add the `-health-sync-containers` flag to `mesh-init` [[GH-36](https://github.com/hashicorp/consul-ecs/pull/36)]
 * Add `-tags`, `-service-name` and `-meta` flags to `mesh-init` [[GH-41](https://github.com/hashicorp/consul-ecs/pull/41)]
 * Add the `-service-name` flag to `health-sync`. [[GH-43](https://github.com/hashicorp/consul-ecs/pull/43)]
+* The ACL controller now reads the Consul service name from the
+  `consul.hashicorp.com/service-name` tag on the ECS task. If the tag
+  does not exist, it uses the Task family as the Consul service name.
+  [[GH-44](https://github.com/hashicorp/consul-ecs/pull/44)]
 
 BREAKING CHANGES
 * `consul-ecs` docker images no longer have the `consul` binary. The